requirements_dev.txt
-r requirements.txt
decorator==4.0.9
flake8==2.5.4
ipdb==0.9.0
ipython==4.1.2
ipython-genutils==0.1.0
mccabe==0.4.0
path.py==8.1.2
pep8==1.7.0
pexpect==4.0.1
pickleshare==0.6
ptyprocess==0.5.1
py==1.4.31
pyflakes==1.1.0
pytest==2.9.1
pytest-django==2.9.1
simplegeneric==0.8.1
traitlets==4.2.1
